Implement _convert_to_grayscale helper function along with tests
We implement the code to convert to grayscale and discover a few things along the way.
First, with three "bands", the number of bands is actually the first dimension of the numpy array.
The function code has been written reflect this, and tests were tweaked as well.

Second, numpy specifies a data type of 8-bit integer, which should be reflected in both the code and the tests.
Now, we're only done testing the numpy shapes. Let's write the test to check how grayscale averaging actually functions.